\name{logicFS-internal}
\alias{minimizePI}
\alias{rm.dom}
\alias{ia.samp}
\alias{cyclic.covering}
\alias{getPerms}
\alias{generateTruthTab}
\alias{getPImps}
\alias{vim.singleBoth}
\alias{vim.singleRemove}
\alias{vim.multiple}
\alias{getNames}
\alias{predict.logregmodel}
\alias{vim.multiple.oneprime}
\alias{check.listprimes}
\alias{check.mat.imp}
\alias{compLarger}
\alias{checkDataCl}

\alias{checkNewTree}
\alias{checkSet}
\alias{compMatImpIndividual1}
\alias{compMatImpIndividual3}
\alias{compMatImpSet1}
\alias{compMatImpSet3}
\alias{correctPreds1}
\alias{correctPredsPermute1}
\alias{correctPredsPermute3}
\alias{correctPredsRemove1}
\alias{correctPredsRemove3}
\alias{correctSetPermute1}
\alias{correctSetPermute3}
\alias{correctSetRemove1}
\alias{correctSetRemove3}
\alias{getKnots}
\alias{getNewTree}
\alias{getTree}
\alias{getVarInTree}
\alias{standardizeMatImp}
\alias{vim.lm}
\alias{vim.lm.oneprime}
\alias{correctPredsPermute2}
\alias{correctPredsRemove2}
\alias{correctSetPermute2}
\alias{correctSetRemove2}

\alias{contr.none}
\alias{contr.snps}
\alias{getXy}
\alias{getXyPred}
\alias{getY}
\alias{makeContrastList}
\alias{modelMat}

\alias{mlogreg.factor}
\alias{predictMLB}
\alias{oobMLR}
\alias{compMatProbMLR}
\alias{vim.MLR}
\alias{pimpMLR}
\alias{vim.MLR}
\alias{getMatPrime}
\alias{getNewProbsMLR}

\alias{vim.singleBoth}
\alias{vim.singleRemove}


\title{Internal logicFS functions}
\description{Internal logicFS functions.}
\details{
   These functions are not meant to be directly called by the user.
}
\author{Holger Schwender, \email{holger.schwender@udo.edu}}
\keyword{internal}{logicFS-internal}
