\name{logicFS-internal}
\alias{vim.singleBoth}
\alias{vim.singleRemove}
\alias{predict.logregmodel}
\alias{contr.none}
\alias{contr.snps}
\alias{getXy}
\alias{triologreg}

\title{Internal logicFS functions}
\description{Internal logicFS functions.}
\details{
   These functions are not meant to be directly called by the user.
}
\author{Holger Schwender, \email{holger.schwender@udo.edu}}
\keyword{internal}
