\name{logicFS-internal}
\alias{minimizePI}
\alias{rm.dom}
\alias{ia.samp}
\alias{cyclic.covering}
\alias{getPerms}
\alias{generateTruthTab}
\alias{getPImps}
\alias{vim.single}
\alias{vim.multiple}
\alias{getNames}
\alias{predict.logregmodel}
\alias{logic.vim2}
\alias{vim.single2}
\alias{vim.multiple.oneprime}
\alias{check.listprimes}
\title{Internal logicFS functions}
\description{Internal logicFS functions.}
\details{
   These functions are not meant to be directly called by the user.
}
\author{Holger Schwender, \email{holger.schwender@udo.edu}}
\keyword{internal}{logicFS-internal}
